+++ upgradetest.go
@@ -130,7 +130,10 @@ func moduleUpgrade(t *testing.T, owner string, repo string, moduleFolderRelative
 	}
 	tmpTestDir := test_structure.CopyTerraformFolderToTemp(t, tmpDirForTag, moduleFolderRelativeToRoot)
 	defer func() {
-		_ = os.RemoveAll(filepath.Clean(tmpTestDir))
+		skip := os.Getenv("SKIP_CLEAN_TMP")
+		if skip == "" {
+			_ = os.RemoveAll(filepath.Clean(tmpTestDir))
+		}
 	}()
 	return diffTwoVersions(t, opts, tmpTestDir, newModulePath)
 }
